requirements.txt
google-cloud-storage>=3.3.0
google-cloud-compute>=1.35.0
google-cloud-functions>=1.20.4
google-auth>=2.32.0
google-auth-oauthlib>=1.2.0
requests>=2.32.5
python-dotenv>=1.0.0
functions-framework>=3.9.2
google-cloud-bigquery>=3.25.0
pytz>=2024.1
